ottixhow/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.7
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect==1.1.0
decorator==5.1.1
dj-database-url==1.0.0
Django==4.0.5
django-bmemcached==0.3.0
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-filter==22.1
django-heroku==0.3.1
django-heroku-memcacheify==1.0.1
django-pylibmc==0.6.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.21.3
executing==0.10.0
filelock==3.7.1
flake8==5.0.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==21.3.0
inflection==0.5.1
install==1.3.5
ipython==8.4.0
itemadapter==0.6.0
itemloaders==1.0.4
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
lxml==4.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
packaging==21.3
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.30
Protego==0.2.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.5
pyflakes==2.5.0
Pygments==2.12.0
pylibmc==1.6.3
pymongo==3.12.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-binary-memcached==0.31.1
pytz==2022.2.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Scrapy==2.6.1
selectorlib==0.16.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.4.0
swagger-spec-validator==2.7.6
tldextract==3.3.0
traitlets==5.3.0
Twisted==22.4.0
typing_extensions==4.2.0
uhashring==2.1
uritemplate==4.1.1
urllib3==1.26.9
w3lib==1.22.0
wcwidth==0.2.5
whitenoise==6.2.0
zipp==3.8.1
zope.interface==5.4.0
